requirements.txt
streamlit
openai
python-dotenv
youtube-transcript-api